backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
mysql-connector-python==8.3.0
firebase-admin==6.5.0
python-dotenv==1.0.1
werkzeug==3.1.3
PyJWT==2.8.0
pusher==3.3.2
bcrypt==4.1.2
flask-limiter==3.5.0

